@@ -19,7 +19,7 @@ import 'dart:io' show Platform;
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
-import 'package:xo/view/choose_player_screen.dart';
+import 'package:xox_madvise/view/choose_player_screen.dart';
 
 // You can also test with your own ad unit IDs by registering your device as a
 // test device. Check the logs for your device's ID value.
@@ -66,26 +66,27 @@ class _AdScreenState extends State<AdScreen> {
 
   void _createInterstitialAd() {
     InterstitialAd.load(
-        adUnitId: Platform.isAndroid
-            ? 'ca-app-pub-3940256099942544/1033173712'
-            : 'ca-app-pub-3940256099942544/4411468910',
-        request: request,
-        adLoadCallback: InterstitialAdLoadCallback(
-          onAdLoaded: (InterstitialAd ad) {
-            print('$ad loaded');
-            _interstitialAd = ad;
-            _numInterstitialLoadAttempts = 0;
-            _interstitialAd!.setImmersiveMode(true);
-          },
-          onAdFailedToLoad: (LoadAdError error) {
-            print('InterstitialAd failed to load: $error.');
-            _numInterstitialLoadAttempts += 1;
-            _interstitialAd = null;
-            if (_numInterstitialLoadAttempts < maxFailedLoadAttempts) {
-              _createInterstitialAd();
-            }
-          },
-        ));
+      adUnitId: Platform.isAndroid
+          ? 'ca-app-pub-3940256099942544/1033173712'
+          : 'ca-app-pub-3940256099942544/4411468910',
+      request: request,
+      adLoadCallback: InterstitialAdLoadCallback(
+        onAdLoaded: (InterstitialAd ad) {
+          print('$ad loaded');
+          _interstitialAd = ad;
+          _numInterstitialLoadAttempts = 0;
+          _interstitialAd!.setImmersiveMode(true);
+        },
+        onAdFailedToLoad: (LoadAdError error) {
+          print('InterstitialAd failed to load: $error.');
+          _numInterstitialLoadAttempts += 1;
+          _interstitialAd = null;
+          if (_numInterstitialLoadAttempts < maxFailedLoadAttempts) {
+            _createInterstitialAd();
+          }
+        },
+      ),
+    );
   }
 
   void _showInterstitialAd() {
@@ -113,25 +114,26 @@ class _AdScreenState extends State<AdScreen> {
 
   void _createRewardedAd() {
     RewardedAd.load(
-        adUnitId: Platform.isAndroid
-            ? 'ca-app-pub-3940256099942544/5224354917'
-            : 'ca-app-pub-3940256099942544/1712485313',
-        request: request,
-        rewardedAdLoadCallback: RewardedAdLoadCallback(
-          onAdLoaded: (RewardedAd ad) {
-            print('$ad loaded.');
-            _rewardedAd = ad;
-            _numRewardedLoadAttempts = 0;
-          },
-          onAdFailedToLoad: (LoadAdError error) {
-            print('RewardedAd failed to load: $error');
-            _rewardedAd = null;
-            _numRewardedLoadAttempts += 1;
-            if (_numRewardedLoadAttempts < maxFailedLoadAttempts) {
-              _createRewardedAd();
-            }
-          },
-        ));
+      adUnitId: Platform.isAndroid
+          ? 'ca-app-pub-3940256099942544/5224354917'
+          : 'ca-app-pub-3940256099942544/1712485313',
+      request: request,
+      rewardedAdLoadCallback: RewardedAdLoadCallback(
+        onAdLoaded: (RewardedAd ad) {
+          print('$ad loaded.');
+          _rewardedAd = ad;
+          _numRewardedLoadAttempts = 0;
+        },
+        onAdFailedToLoad: (LoadAdError error) {
+          print('RewardedAd failed to load: $error');
+          _rewardedAd = null;
+          _numRewardedLoadAttempts += 1;
+          if (_numRewardedLoadAttempts < maxFailedLoadAttempts) {
+            _createRewardedAd();
+          }
+        },
+      ),
+    );
   }
 
   void _showRewardedAd() {
@@ -156,33 +158,35 @@ class _AdScreenState extends State<AdScreen> {
 
     _rewardedAd!.setImmersiveMode(true);
     _rewardedAd!.show(
-        onUserEarnedReward: (AdWithoutView ad, RewardItem reward) {
-      print('$ad with reward $RewardItem(${reward.amount}, ${reward.type})');
-    });
+      onUserEarnedReward: (AdWithoutView ad, RewardItem reward) {
+        print('$ad with reward $RewardItem(${reward.amount}, ${reward.type})');
+      },
+    );
     _rewardedAd = null;
   }
 
   void _createRewardedInterstitialAd() {
     RewardedInterstitialAd.load(
-        adUnitId: Platform.isAndroid
-            ? 'ca-app-pub-3940256099942544/5354046379'
-            : 'ca-app-pub-3940256099942544/6978759866',
-        request: request,
-        rewardedInterstitialAdLoadCallback: RewardedInterstitialAdLoadCallback(
-          onAdLoaded: (RewardedInterstitialAd ad) {
-            print('$ad loaded.');
-            _rewardedInterstitialAd = ad;
-            _numRewardedInterstitialLoadAttempts = 0;
-          },
-          onAdFailedToLoad: (LoadAdError error) {
-            print('RewardedInterstitialAd failed to load: $error');
-            _rewardedInterstitialAd = null;
-            _numRewardedInterstitialLoadAttempts += 1;
-            if (_numRewardedInterstitialLoadAttempts < maxFailedLoadAttempts) {
-              _createRewardedInterstitialAd();
-            }
-          },
-        ));
+      adUnitId: Platform.isAndroid
+          ? 'ca-app-pub-3940256099942544/5354046379'
+          : 'ca-app-pub-3940256099942544/6978759866',
+      request: request,
+      rewardedInterstitialAdLoadCallback: RewardedInterstitialAdLoadCallback(
+        onAdLoaded: (RewardedInterstitialAd ad) {
+          print('$ad loaded.');
+          _rewardedInterstitialAd = ad;
+          _numRewardedInterstitialLoadAttempts = 0;
+        },
+        onAdFailedToLoad: (LoadAdError error) {
+          print('RewardedInterstitialAd failed to load: $error');
+          _rewardedInterstitialAd = null;
+          _numRewardedInterstitialLoadAttempts += 1;
+          if (_numRewardedInterstitialLoadAttempts < maxFailedLoadAttempts) {
+            _createRewardedInterstitialAd();
+          }
+        },
+      ),
+    );
   }
 
   void _showRewardedInterstitialAd() {
@@ -192,26 +196,27 @@ class _AdScreenState extends State<AdScreen> {
     }
     _rewardedInterstitialAd!.fullScreenContentCallback =
         FullScreenContentCallback(
-      onAdShowedFullScreenContent: (RewardedInterstitialAd ad) =>
-          print('$ad onAdShowedFullScreenContent.'),
-      onAdDismissedFullScreenContent: (RewardedInterstitialAd ad) {
-        print('$ad onAdDismissedFullScreenContent.');
-        ad.dispose();
-        _createRewardedInterstitialAd();
-      },
-      onAdFailedToShowFullScreenContent:
-          (RewardedInterstitialAd ad, AdError error) {
-        print('$ad onAdFailedToShowFullScreenContent: $error');
-        ad.dispose();
-        _createRewardedInterstitialAd();
-      },
-    );
+          onAdShowedFullScreenContent: (RewardedInterstitialAd ad) =>
+              print('$ad onAdShowedFullScreenContent.'),
+          onAdDismissedFullScreenContent: (RewardedInterstitialAd ad) {
+            print('$ad onAdDismissedFullScreenContent.');
+            ad.dispose();
+            _createRewardedInterstitialAd();
+          },
+          onAdFailedToShowFullScreenContent:
+              (RewardedInterstitialAd ad, AdError error) {
+                print('$ad onAdFailedToShowFullScreenContent: $error');
+                ad.dispose();
+                _createRewardedInterstitialAd();
+              },
+        );
 
     _rewardedInterstitialAd!.setImmersiveMode(true);
     _rewardedInterstitialAd!.show(
-        onUserEarnedReward: (AdWithoutView ad, RewardItem reward) {
-      print('$ad with reward $RewardItem(${reward.amount}, ${reward.type})');
-    });
+      onUserEarnedReward: (AdWithoutView ad, RewardItem reward) {
+        print('$ad with reward $RewardItem(${reward.amount}, ${reward.type})');
+      },
+    );
     _rewardedInterstitialAd = null;
   }
 
@@ -226,107 +231,116 @@ class _AdScreenState extends State<AdScreen> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: Builder(builder: (BuildContext context) {
-        return Scaffold(
-          appBar: AppBar(
-            title: const Text('AdMob Plugin example app'),
-            actions: <Widget>[
-              PopupMenuButton<String>(
-                onSelected: (String result) {
-                  switch (result) {
-                    case interstitialButtonText:
-                      _showInterstitialAd();
-                      break;
-                    case rewardedButtonText:
-                      _showRewardedAd();
-                      break;
-                    case rewardedInterstitialButtonText:
-                      _showRewardedInterstitialAd();
-                      break;
-                    case fluidButtonText:
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => ChoosePlayerScreen()),
-                      );
-                      break;
-                    case inlineAdaptiveButtonText:
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => ChoosePlayerScreen()),
-                      );
-                      break;
-                    case anchoredAdaptiveButtonText:
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => ChoosePlayerScreen()),
-                      );
-                      break;
-                    case nativeTemplateButtonText:
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => ChoosePlayerScreen()),
-                      );
-                      break;
-                    case webviewExampleButtonText:
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => ChoosePlayerScreen()),
-                      );
-                      break;
-                    default:
-                      throw AssertionError('unexpected button: $result');
-                  }
-                },
-                itemBuilder: (BuildContext context) => <PopupMenuEntry<String>>[
-                  PopupMenuItem<String>(
-                    value: interstitialButtonText,
-                    child: Text(interstitialButtonText),
-                  ),
-                  PopupMenuItem<String>(
-                    value: rewardedButtonText,
-                    child: Text(rewardedButtonText),
-                  ),
-                  PopupMenuItem<String>(
-                    value: rewardedInterstitialButtonText,
-                    child: Text(rewardedInterstitialButtonText),
-                  ),
-                  PopupMenuItem<String>(
-                    value: fluidButtonText,
-                    child: Text(fluidButtonText),
-                  ),
-                  PopupMenuItem<String>(
-                    value: inlineAdaptiveButtonText,
-                    child: Text(inlineAdaptiveButtonText),
-                  ),
-                  PopupMenuItem<String>(
-                    value: anchoredAdaptiveButtonText,
-                    child: Text(anchoredAdaptiveButtonText),
-                  ),
-                  PopupMenuItem<String>(
-                    value: nativeTemplateButtonText,
-                    child: Text(nativeTemplateButtonText),
-                  ),
-                  PopupMenuItem<String>(
-                    value: webviewExampleButtonText,
-                    child: Text(webviewExampleButtonText),
-                  ),
-                ],
-              ),
-            ],
-          ),
-          body: SafeArea(
+      home: Builder(
+        builder: (BuildContext context) {
+          return Scaffold(
+            appBar: AppBar(
+              title: const Text('AdMob Plugin example app'),
+              actions: <Widget>[
+                PopupMenuButton<String>(
+                  onSelected: (String result) {
+                    switch (result) {
+                      case interstitialButtonText:
+                        _showInterstitialAd();
+                        break;
+                      case rewardedButtonText:
+                        _showRewardedAd();
+                        break;
+                      case rewardedInterstitialButtonText:
+                        _showRewardedInterstitialAd();
+                        break;
+                      case fluidButtonText:
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => ChoosePlayerScreen(),
+                          ),
+                        );
+                        break;
+                      case inlineAdaptiveButtonText:
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => ChoosePlayerScreen(),
+                          ),
+                        );
+                        break;
+                      case anchoredAdaptiveButtonText:
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => ChoosePlayerScreen(),
+                          ),
+                        );
+                        break;
+                      case nativeTemplateButtonText:
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => ChoosePlayerScreen(),
+                          ),
+                        );
+                        break;
+                      case webviewExampleButtonText:
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => ChoosePlayerScreen(),
+                          ),
+                        );
+                        break;
+                      default:
+                        throw AssertionError('unexpected button: $result');
+                    }
+                  },
+                  itemBuilder: (BuildContext context) =>
+                      <PopupMenuEntry<String>>[
+                        PopupMenuItem<String>(
+                          value: interstitialButtonText,
+                          child: Text(interstitialButtonText),
+                        ),
+                        PopupMenuItem<String>(
+                          value: rewardedButtonText,
+                          child: Text(rewardedButtonText),
+                        ),
+                        PopupMenuItem<String>(
+                          value: rewardedInterstitialButtonText,
+                          child: Text(rewardedInterstitialButtonText),
+                        ),
+                        PopupMenuItem<String>(
+                          value: fluidButtonText,
+                          child: Text(fluidButtonText),
+                        ),
+                        PopupMenuItem<String>(
+                          value: inlineAdaptiveButtonText,
+                          child: Text(inlineAdaptiveButtonText),
+                        ),
+                        PopupMenuItem<String>(
+                          value: anchoredAdaptiveButtonText,
+                          child: Text(anchoredAdaptiveButtonText),
+                        ),
+                        PopupMenuItem<String>(
+                          value: nativeTemplateButtonText,
+                          child: Text(nativeTemplateButtonText),
+                        ),
+                        PopupMenuItem<String>(
+                          value: webviewExampleButtonText,
+                          child: Text(webviewExampleButtonText),
+                        ),
+                      ],
+                ),
+              ],
+            ),
+            body: SafeArea(
               child: Container(
-            height: Get.height,
-            width: Get.width,
-            color: Colors.red,
-          )),
-        );
-      }),
+                height: Get.height,
+                width: Get.width,
+                color: Colors.red,
+              ),
+            ),
+          );
+        },
+      ),
     );
   }
 }
